backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.3.2
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
slowapi==0.1.9
python-multipart==0.0.6
asyncio==3.4.3
APScheduler==3.10.4
gspread==6.0.0
google-auth==2.26.2
